feat(quotations): exportar el PDF de la cotizacion
`POST /api/v1/tenants/{tenantId}/quotations/{quotationId}/pdf` genera el PDF si
hace falta y devuelve una URL firmada contra el bucket privado. Devuelve el
enlace y no los bytes: la descarga va directo de R2 al navegador, sin que el
documento pase por la API.

Solo regenera cuando `Quotation.Version` avanzo. Generar cuesta una llamada de
red a `qcode-pdf` mas una subida a R2, y la mayoria de las exportaciones son de
cotizaciones que nadie toco desde la anterior: sin esa comparacion, abrir dos
veces la misma pantalla pagaria las dos. Es el motivo de existir de
`quotation_pdfs`, y hay una prueba que lo fija.

El documento se arma desde la misma `QuotationResponse` que dibuja la pantalla,
no desde el agregado: `QuotationResponseComposer` ya resolvio el cliente, la
empresa de la cuenta de cobro y el nombre de cada producto. Si el PDF y la
pantalla se armaran por caminos distintos, terminarian diciendo cosas distintas
de la misma cotizacion y la diferencia la descubriria el cliente.

Pide `QuotationRead` y no `QuotationManage`, mismo criterio que el export de
clientes: exportar es leer la cotizacion en otro formato, y quien la ve en
pantalla ya ve sus precios.

El PDF no se registra como `FileResource`. Esa maquina de estados --pendiente,
subido, escaneado, disponible-- existe para archivos que sube un cliente no
confiable: verifica tamaño, calcula checksum y pasa por antivirus. Un PDF que
genera el propio backend no necesita nada de eso.

`quotation_version` es `bigint`: `Quotation.Version` es `long`. La migracion
anterior lo tenia como `integer` y se regenero.

+++ src/Modules/Quotations/Modules.Quotations.Domain/QuotationPdf.cs
@@ -20,7 +20,7 @@ public sealed class QuotationPdf
         QuotationId quotationId,
         Guid tenantId,
         string storageKey,
-        int quotationVersion,
+        long quotationVersion,
         DateTimeOffset generatedAt)
     {
         QuotationId = quotationId;
@@ -40,7 +40,7 @@ public sealed class QuotationPdf
 
     /// <summary><c>Quotation.Version</c> al momento de generar. No hace falta un mecanismo
     /// propio de invalidación: el agregado ya lo incrementa en cada cambio.</summary>
-    public int QuotationVersion { get; private set; }
+    public long QuotationVersion { get; private set; }
 
     public DateTimeOffset GeneratedAt { get; private set; }
 
@@ -48,7 +48,7 @@ public sealed class QuotationPdf
         QuotationId quotationId,
         Guid tenantId,
         string storageKey,
-        int quotationVersion,
+        long quotationVersion,
         DateTimeOffset generatedAt) =>
         new(quotationId, tenantId, storageKey, quotationVersion, generatedAt);
 
@@ -57,9 +57,9 @@ public sealed class QuotationPdf
     /// —una restauración de base, una escritura fuera de orden— no se arregla regenerándolo, y
     /// tratarlo como obsoleto lo haría regenerar en cada pedido, para siempre.
     /// </summary>
-    public bool IsStaleFor(int quotationVersion) => QuotationVersion < quotationVersion;
+    public bool IsStaleFor(long quotationVersion) => QuotationVersion < quotationVersion;
 
-    public void Regenerate(string storageKey, int quotationVersion, DateTimeOffset generatedAt)
+    public void Regenerate(string storageKey, long quotationVersion, DateTimeOffset generatedAt)
     {
         StorageKey = storageKey;
         QuotationVersion = quotationVersion;
